+++ internal/query/animefiles_test.go
@@ -43,7 +43,7 @@ func TestGetAnimeFiles(t *testing.T) {
 	if err := InsertAnime(db, a); err != nil {
 		t.Fatalf("Error inserting anime: %s", err)
 	}
-	efs := []EpisodeFile{{EpisodeID: 1, Path: "/foobar"}}
+	efs := []EpisodeFile{{EpisodeID: 1, EID: 113, Path: "/foobar"}}
 	if err := InsertEpisodeFiles(db, efs); err != nil {
 		t.Fatalf("Error inserting episode file: %s", err)
 	}
@@ -63,7 +63,7 @@ func TestGetAnimeFiles(t *testing.T) {
 				Length: 25,
 			},
 			Files: []EpisodeFile{
-				{EpisodeID: 1, Path: "/foobar"},
+				{EpisodeID: 1, EID: 113, Path: "/foobar"},
 			},
 		},
 	}